Separated Models, Interfaces, ViewModels (and a lot more) into Base class library
