+++ ApiExample.Base/Attributes/ValidImageAttribute.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace ApiExample.Attributes
+namespace ApiExample.Base.Attributes
 {
     public class ValidImageAttribute : ValidationAttribute
     {
