+++ ApiExample.Base/Interfaces/IHasTextContent.cs
@@ -1,4 +1,4 @@
-﻿namespace ApiExample.Interfaces
+﻿namespace ApiExample.Base.Interfaces
 {
     public interface IHasTextContent
     {
