+++ ApiExample.Base/Models/Publication.cs
@@ -2,9 +2,9 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Diagnostics.CodeAnalysis;
-using ApiExample.Interfaces;
+using ApiExample.Base.Interfaces;
 
-namespace ApiExample.Models
+namespace ApiExample.Base.Models
 {
     public class Publication : PostedBase, IHasTextContent
     {
