+++ ApiExample.Base/Repositories/ICommentsRepository.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using ApiExample.Models;
+using ApiExample.Base.Models;
 
-namespace ApiExample.Repositories
+namespace ApiExample.Base.Repositories
 {
     public interface ICommentsRepository
     {
