+++ ApiExample.Base/Repositories/PublicationsRepository.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using ApiExample.DbContexts;
-using ApiExample.Models;
+using ApiExample.Base.DbContexts;
+using ApiExample.Base.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace ApiExample.Repositories
+namespace ApiExample.Base.Repositories
 {
     public class PublicationsRepository : IPublicationsRepository
     {
