+++ ApiExample.Base/ViewModels/Comments/CreateCommentVM.cs
@@ -1,38 +1,30 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using ApiExample.Attributes;
-using ApiExample.Models;
+using ApiExample.Base.Models;
 
-namespace ApiExample.ViewModels.Publications
+namespace ApiExample.ViewModels.Comments
 {
-    public class CreatePublicationVM
+    public class CreateCommentVM
     {
         #region Public Properties
 
-        [StringLength(300)]
+        [Required]
         public string Content { get; set; }
 
-        [ValidImage]
-        [MaxLength(1 << 22, ErrorMessage = "~4 MB is the limit for image upload :(")]
-        public byte[] Image { get; set; } = null;
-
         [Required]
-        [StringLength(50)]
         public string Pseudonym { get; set; } = "Anon";
 
         #endregion Public Properties
 
         #region Public Methods
 
-        public Publication ToModel()
+        public Comment ToModel(string publicationId)
         {
-            return new Publication()
+            return new Comment()
             {
                 AuthorPseudonym = Pseudonym,
                 Content = Content,
-                Image = Image,
-                Comments = new List<Comment>(),
+                PublicationId = publicationId,
                 PublicationTime = DateTime.UtcNow,
                 Rating = 0
             };
