+++ ApiExample.Base/ViewModels/Publications/CreatePublicationVM.cs
@@ -1,30 +1,38 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using ApiExample.Models;
+using ApiExample.Base.Attributes;
+using ApiExample.Base.Models;
 
-namespace ApiExample.ViewModels.Comments
+namespace ApiExample.ViewModels.Publications
 {
-    public class CreateCommentVM
+    public class CreatePublicationVM
     {
         #region Public Properties
 
-        [Required]
+        [StringLength(300)]
         public string Content { get; set; }
 
+        [ValidImage]
+        [MaxLength(1 << 22, ErrorMessage = "~4 MB is the limit for image upload :(")]
+        public byte[] Image { get; set; } = null;
+
         [Required]
+        [StringLength(50)]
         public string Pseudonym { get; set; } = "Anon";
 
         #endregion Public Properties
 
         #region Public Methods
 
-        public Comment ToModel(string publicationId)
+        public Publication ToModel()
         {
-            return new Comment()
+            return new Publication()
             {
                 AuthorPseudonym = Pseudonym,
                 Content = Content,
-                PublicationId = publicationId,
+                Image = Image,
+                Comments = new List<Comment>(),
                 PublicationTime = DateTime.UtcNow,
                 Rating = 0
             };
