+++ ApiExample/ApiExample.Chat.csproj
@@ -17,4 +17,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ApiExample.Base\ApiExample.Base.csproj" />
+  </ItemGroup>
+
 </Project>
