+++ ApiExample/Controllers/PublicationsController.cs
@@ -1,14 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using ApiExample.Attributes;
-using ApiExample.Models;
-using ApiExample.Repositories;
+using ApiExample.Base.Attributes;
+using ApiExample.Base.Models;
+using ApiExample.Base.Repositories;
 using ApiExample.ViewModels.Comments;
 using ApiExample.ViewModels.Publications;
 using Microsoft.AspNetCore.Mvc;
 
-namespace ApiExample.Controllers
+namespace ApiExample.Chat.Controllers
 {
     [ApiController]
     [Route("[controller]")]
